+++ includes/math/perlin/spk_perlin_3D.hpp
@@ -63,7 +63,11 @@ namespace spk
 		template<typename TVectorType>
 		float sample(const spk::IVector3<TVectorType>& p_position) const
 		{
-			return (sample(p_position.x, p_position.y, p_position.z));
+			return (sample(
+					static_cast<float>(p_position.x),
+					static_cast<float>(p_position.y),
+					static_cast<float>(p_position.z))
+				);
 		}
 	};
 }
